code/requirements.txt
awesome-slugify
beautifulsoup4
requests

flask==1.0.3
flask-cors==3.0.7
flask-caching
praw==6.2.0
pymongo==3.8.0
python-dateutil==2.8.0
xmltodict==0.12.0
